Add ClusterRole and other related objects to ClusterOperator relatedObjects
To facilitate the gathering of resources relevant to cluster-csi-snapshot-controller-operator,
the ClusterOperator's `relatedObjects` is expanded to reference ClusterRole and other resources
that might be relevant to debugging the component, as described in the docs
(https://github.com/openshift/enhancements/blob/master/dev-guide/cluster-version-operator/dev/clusteroperator.md#related-objects).
This commit also add entries to ClusterOperator manifest, so the CVO can put that entry in place if the operator fails to install.